87 urgent datalab registration form please add option to indicate institutional affiliation and institutional url (#89)
This PR introduces institutional affiliation and URL fields to the registration and profile flows, updates plan metadata, and brings in updated styling and modal components. The affiliation fields can be hidden to cope with older version of IML.

fix #87 - there is a flag `hideAffiliationFields` on the ProfileForm included into the SignUpModal component that enable/ disable the affiliation field display and validation. This allow us to use the old version of IML

Components now come from the `impresso-ui-components` repository and they are wrapped inside Vue Modal
As a result, several react component can be dismissed.

---------

Co-authored-by: Daniele Guido <1181642+danieleguido@users.noreply.github.com>